+++ src/Installer/Database/PostgresDatabase.php
@@ -26,8 +26,9 @@ class PostgresDatabase extends AbstractDatabase
     public function getCreateDatabaseCommand()
     {
         $user     = escapeshellarg($this->user);
-        $createDB = escapeshellarg(sprintf('CREATE DATABASE %s;', $this->name));
+        $host     = escapeshellarg($this->host);
+        $createDB = escapeshellarg(sprintf('CREATE DATABASE "%s";', $this->name));
 
-        return sprintf('psql -c %s -U %s', $createDB, $user);
+        return sprintf('psql -c %s -U %s -h %s', $createDB, $user, $host);
     }
 }