+++ tests/Installer/Database/PostgresDatabaseTest.php
@@ -28,7 +28,7 @@ class PostgresDatabaseTest extends \PHPUnit_Framework_TestCase
         $database->pass = 'TestPass';
         $database->host = 'TestHost';
 
-        $expected = 'psql -c \'CREATE DATABASE TestName;\' -U \'TestUser\'';
+        $expected = 'psql -c \'CREATE DATABASE "TestName";\' -U \'TestUser\' -h \'TestHost\'';
         $this->assertEquals($expected, $database->getCreateDatabaseCommand());
     }
 }