+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 The format of this change log follows the advice given at [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+### Fixed
+- Pass host when connecting with MySQL and Postgres.
+- Add quotes around database name for MySQL and Postgres.
 
 ## [1.1.0] - 2015-10-19
 ### Added